refactor - extract metadata type and use it in scanner
